Refactoring Write Repository to follow DRY principle
